server.js분리
